# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Daedalus API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/questions/{question_id}/attempts": {
      "post": {
        "tags": [
          "grading"
        ],
        "summary": "Answer Question",
        "description": "Answer a question, have the answer graded, and reschedule the question. Only questions\nin the library can be answered: one that was rejected or retired is not there to practise.",
        "operationId": "answer_question",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "grading"
        ],
        "summary": "List Attempts",
        "description": "Your answers to a question, newest first.",
        "operationId": "list_attempts",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttemptOut"
                  },
                  "title": "Response List Attempts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/attempts/{attempt_id}/grades": {
      "post": {
        "tags": [
          "grading"
        ],
        "summary": "Grade Again",
        "description": "Grade an attempt again, after a failed grade or with a changed grader. Every earlier\ngrade is kept. The first successful grade reschedules the question; later ones don't.",
        "operationId": "grade_again",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/attempts/{attempt_id}": {
      "get": {
        "tags": [
          "grading"
        ],
        "summary": "Get Attempt",
        "description": "One of your attempts with every grade it was given.",
        "operationId": "get_attempt",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerIn": {
        "properties": {
          "answer": {
            "type": "string",
            "maxLength": 8000,
            "title": "Answer"
          },
          "seconds": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 86400.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Seconds"
          },
          "time_limit": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3600.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "AnswerIn"
      },
...
      "AttemptOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seconds"
          },
          "time_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "grades": {
            "items": {
              "$ref": "#/components/schemas/GradeOut"
            },
            "type": "array",
            "title": "Grades"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "earned": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EarnedOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_id",
          "question",
          "answer",
          "seconds",
          "time_limit",
          "created_at",
          "grades",
          "review",
          "earned"
        ],
        "title": "AttemptOut"
      },
...
      "ClaimOut": {
        "properties": {
          "claim": {
            "type": "string",
            "title": "Claim"
          },
          "verdict": {
            "type": "string",
            "title": "Verdict"
          },
          "why": {
            "type": "string",
            "title": "Why"
          },
          "chunk_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id"
          },
          "citation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citation"
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link"
          }
        },
        "type": "object",
        "required": [
          "claim",
          "verdict",
          "why",
          "chunk_id",
          "citation",
          "link"
        ],
        "title": "ClaimOut"
      },
...
      "EarnedOut": {
        "properties": {
          "xp": {
            "type": "integer",
            "title": "Xp"
          },
          "parts": {
            "$ref": "#/components/schemas/XpOut"
          },
          "total_xp": {
            "type": "integer",
            "title": "Total Xp"
          },
          "level": {
            "$ref": "#/components/schemas/LevelOut"
          },
          "level_up": {
            "type": "boolean",
            "title": "Level Up"
          },
          "streak": {
            "type": "integer",
            "title": "Streak"
          },
          "coins": {
            "items": {
              "$ref": "#/components/schemas/CoinOut"
            },
            "type": "array",
            "title": "Coins"
          }
        },
        "type": "object",
        "required": [
          "xp",
          "parts",
          "total_xp",
          "level",
          "level_up",
          "streak",
          "coins"
        ],
        "title": "EarnedOut"
      },
...
      "GradeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "grader_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grader Model"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "coverage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "contradicted": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contradicted"
          },
          "clarity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clarity"
          },
          "key_points": {
            "items": {
              "$ref": "#/components/schemas/KeyPointGradeOut"
            },
            "type": "array",
            "title": "Key Points"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/ClaimOut"
            },
            "type": "array",
            "title": "Claims"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Gaps"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "improved_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Improved Answer"
          },
          "follow_up": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up"
          },
          "usage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Usage"
          },
          "seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seconds"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "rating": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RatingOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "error",
          "grader_model",
          "prompt_version",
          "score",
          "coverage",
          "contradicted",
          "clarity",
          "key_points",
          "claims",
          "strengths",
          "gaps",
          "errors",
          "improved_answer",
          "follow_up",
          "usage",
          "seconds",
          "created_at",
          "rating"
        ],
        "title": "GradeOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeyPointGradeOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "weight": {
            "type": "integer",
            "title": "Weight"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "answer_quote": {
            "type": "string",
            "title": "Answer Quote"
          },
          "quote_found": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Found"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "weight",
          "status",
          "answer_quote",
          "quote_found"
        ],
        "title": "KeyPointGradeOut"
      },
...
      "LevelOut": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "of": {
            "type": "integer",
            "title": "Of"
          },
          "start": {
            "type": "integer",
            "title": "Start"
          },
          "next_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Start"
          },
          "next_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Name"
          }
        },
        "type": "object",
        "required": [
          "number",
          "name",
          "of",
          "start",
          "next_start",
          "next_name"
        ],
        "title": "LevelOut"
      },
...
      "RatingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "question_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          },
          "grade_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Id"
          },
          "value": {
            "type": "integer",
            "enum": [
              1,
              -1
            ],
            "title": "Value"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_id",
          "grade_id",
          "value",
          "note",
          "created_at"
        ],
        "title": "RatingOut"
      },
      "ReviewOut": {
        "properties": {
          "rating": {
            "type": "string",
            "enum": [
              "again",
              "hard",
              "good",
              "easy"
            ],
            "title": "Rating"
          },
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "due": {
            "type": "string",
            "format": "date",
            "title": "Due"
          },
          "interval": {
            "type": "integer",
            "title": "Interval"
          }
        },
        "type": "object",
        "required": [
          "rating",
          "day",
          "due",
          "interval"
        ],
        "title": "ReviewOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "XpOut": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "answered": {
            "type": "integer",
            "title": "Answered"
          },
          "due": {
            "type": "integer",
            "title": "Due"
          },
          "interview": {
            "type": "integer",
            "title": "Interview"
          },
          "streak": {
            "type": "integer",
            "title": "Streak"
          }
        },
        "type": "object",
        "required": [
          "score",
          "answered",
          "due",
          "interview",
          "streak"
        ],
        "title": "XpOut",
        "description": "The XP an answer earned, part by part: ten times its score, five for answering, half\nas much again for a question that was due, five inside an interview limit, and, on the\nday's first answer, the streak's length that day, up to ten."
      }
    }
  }
}